+++ rmk/tests/keyboard_combo_test.rs
@@ -55,7 +55,7 @@ pub fn get_combos_config() -> CombosConfig {
 
 mod combo_test {
     use embassy_futures::block_on;
-    use rmk::config::{BehaviorConfig, OneShotConfig, TapHoldConfig};
+    use rmk::config::{BehaviorConfig, OneShotConfig, TapHoldConfig, TapHoldMode};
     use rmk::keycode::KeyCode;
     use rmk::th;
     use rusty_fork::rusty_fork_test;
@@ -184,9 +184,8 @@ mod combo_test {
                     let behavior_config = BehaviorConfig {
                         tap_hold: TapHoldConfig {
                             enable_hrm: true,
-                            permissive_hold: true,
+                            mode: TapHoldMode::PermissiveHold,
                             chordal_hold: false,
-                            post_wait_time: Duration::from_millis(0),
                             ..TapHoldConfig::default()
                         },
                         combo: CombosConfig {