+++ rmk/tests/keyboard_tapdance_test.rs
@@ -33,7 +33,7 @@ fn create_tap_dance_config() -> TapDancesConfig {
 
 mod tap_dance_test {
     use embassy_futures::block_on;
-    use rmk::config::TapHoldConfig;
+    use rmk::config::{TapHoldConfig, TapHoldMode};
     use rmk::keyboard::Keyboard;
     use rusty_fork::rusty_fork_test;
 
@@ -51,7 +51,7 @@ mod tap_dance_test {
             tap_dance: create_tap_dance_config(),
             tap_hold: TapHoldConfig {
                 enable_hrm: true,
-                permissive_hold: true,
+                mode: TapHoldMode::PermissiveHold,
                 ..Default::default()
             },
             ..Default::default()