+++ rmk/tests/keyboard_taphold_test.rs
@@ -1,15 +1,13 @@
 pub mod common;
 
 use embassy_time::Duration;
-use rmk::config::TapHoldConfig;
+use rmk::config::{TapHoldConfig, TapHoldMode};
 
 fn tap_hold_config_with_hrm_and_permissive_hold() -> TapHoldConfig {
     TapHoldConfig {
         enable_hrm: true,
-        permissive_hold: true,
+        mode: TapHoldMode::PermissiveHold,
         chordal_hold: false,
-        hold_on_other_press: false,
-        post_wait_time: Duration::from_millis(0),
         ..TapHoldConfig::default()
     }
 }
@@ -18,9 +16,7 @@ fn tap_hold_config_with_hrm_and_chordal_hold() -> TapHoldConfig {
     TapHoldConfig {
         enable_hrm: true,
         chordal_hold: true,
-        permissive_hold: true,
-        hold_on_other_press: false,
-        post_wait_time: Duration::from_millis(0),
+        mode: TapHoldMode::PermissiveHold,
         ..TapHoldConfig::default()
     }
 }
@@ -125,8 +121,7 @@ mod tap_hold_test {
             let config =BehaviorConfig {
                 tap_hold: TapHoldConfig {
                     enable_hrm: true,
-                    permissive_hold: true,
-                    post_wait_time: Duration::from_millis(0),
+                    mode: TapHoldMode::PermissiveHold,
                     ..Default::default()
                 },
                 ..Default::default()
@@ -161,8 +156,7 @@ mod tap_hold_test {
             let config = BehaviorConfig {
                 tap_hold: TapHoldConfig {
                     enable_hrm: true,
-                    permissive_hold: true,
-                    post_wait_time: Duration::from_millis(0),
+                    mode: TapHoldMode::PermissiveHold,
                     ..Default::default()
                 },
                 ..Default::default()
@@ -741,5 +735,32 @@ mod tap_hold_test {
             };
         }
 
+        // Ref: https://github.com/HaoboGu/rmk/pull/496
+        #[test]
+        fn test_taphold_with_previous_rolling_keypress() {
+            key_sequence_test! {
+                keyboard: {
+                    let behavior_config = BehaviorConfig {
+                        tap_hold: tap_hold_config_with_hrm_and_permissive_hold(),
+                        ..BehaviorConfig::default()
+                    };
+                    create_test_keyboard_with_config(behavior_config)
+                },
+                sequence: [
+                    [1, 1, true, 20],  // Press Q
+                    [4, 5, true, 150],  // Press lt!(1, space)
+                    [1, 1, false, 20],  // Release Q
+                    [1, 3, true, 10], // Press E(on layer 1 is W)
+                    [1, 3, false, 10], // Release E(should trigger W)
+                    [4, 5, false, 10], // Release lt!(1, space)
+                ],
+                expected_reports: [
+                    [0, [kc_to_u8!(Q), 0, 0, 0, 0, 0]], // Tap Q
+                    [0, [0, 0, 0, 0, 0, 0]], // All released
+                    [0, [kc_to_u8!(W), 0, 0, 0, 0, 0]], // Tap W
+                    [0, [0, 0, 0, 0, 0, 0]], // All released
+                ]
+            };
+        }
     }
 }